+++ src/test/java/org/cryptomator/linux/keychain/SecretServiceKeychainAccessTest.java
@@ -17,17 +17,17 @@ import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 /**
- * Unit tests for GNOME keyring access via DBUS.
+ * Unit tests for Secret Service access via Dbus.
  */
-@EnabledIfEnvironmentVariable(named = "DISPLAY", matches = ".*")
-public class GnomeKeyringKeychainAccessTest {
+@EnabledIfEnvironmentVariable(named = "DBUS_SESSION_BUS_ADDRESS", matches = ".*")
+public class SecretServiceKeychainAccessTest {
 
 	private static boolean isInstalled;
 
 	@BeforeAll
 	public static void checkSystemAndSetup() throws IOException {
 		ProcessBuilder dbusSend = new ProcessBuilder("dbus-send", "--print-reply", "--dest=org.freedesktop.DBus", "/org/freedesktop/DBus", "org.freedesktop.DBus.ListNames");
-		ProcessBuilder grep = new ProcessBuilder("grep", "-q", "org.gnome.keyring");
+		ProcessBuilder grep = new ProcessBuilder("grep", "-q", "org.freedesktop.secrets");
 		try {
 			Process end = ProcessBuilder.startPipeline(List.of(dbusSend, grep)).get(1);
 			if (end.waitFor(1000, TimeUnit.MILLISECONDS)) {
@@ -42,28 +42,29 @@ public class GnomeKeyringKeychainAccessTest {
 
 	@Test
 	public void testIsSupported() {
-		var gnomeKeyring = new GnomeKeyringKeychainAccess();
-		Assertions.assertEquals(isInstalled, gnomeKeyring.isSupported());
+		var service = new SecretServiceKeychainAccess();
+		Assertions.assertEquals(isInstalled, service.isSupported());
 	}
 
 	@Nested
 	@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-	@EnabledIf("gnomeKeyringAvailableAndUnlocked")
+	@EnabledIf("serviceAvailableAndUnlocked")
 	class FunctionalTests {
 
 		static final String KEY_ID = "cryptomator-test-" + UUID.randomUUID();
-		final GnomeKeyringKeychainAccess gnomeKeyring = new GnomeKeyringKeychainAccess();
+		final static SecretServiceKeychainAccess KEYRING = new SecretServiceKeychainAccess();
 
 		@Test
 		@Order(1)
 		public void testStore() throws KeychainAccessException {
-			gnomeKeyring.storePassphrase(KEY_ID, "cryptomator-test", "p0ssw0rd");
+			KEYRING.isSupported(); // ensure encrypted session
+			KEYRING.storePassphrase(KEY_ID, "cryptomator-test", "p0ssw0rd");
 		}
 
 		@Test
 		@Order(2)
 		public void testLoad() throws KeychainAccessException {
-			var passphrase = gnomeKeyring.loadPassphrase(KEY_ID);
+			var passphrase = KEYRING.loadPassphrase(KEY_ID);
 			Assertions.assertNotNull(passphrase);
 			Assertions.assertEquals("p0ssw0rd", String.copyValueOf(passphrase));
 		}
@@ -71,19 +72,19 @@ public class GnomeKeyringKeychainAccessTest {
 		@Test
 		@Order(3)
 		public void testDelete() throws KeychainAccessException {
-			gnomeKeyring.deletePassphrase(KEY_ID);
+			KEYRING.deletePassphrase(KEY_ID);
 		}
 
 		@Test
 		@Order(4)
 		public void testLoadNotExisting() throws KeychainAccessException {
-			var result = gnomeKeyring.loadPassphrase(KEY_ID);
+			var result = KEYRING.loadPassphrase(KEY_ID);
 			Assertions.assertNull(result);
 		}
 
-		public static boolean gnomeKeyringAvailableAndUnlocked() {
-			var secretServiceKeychain = new GnomeKeyringKeychainAccess();
-			return secretServiceKeychain.isSupported() && !secretServiceKeychain.isLocked();
+		public static boolean serviceAvailableAndUnlocked() {
+			var service = new SecretServiceKeychainAccess();
+			return service.isSupported() && !service.isLocked();
 		}
 	}